Found test that doesn't work well.
I'm not sure how we'd like the students to be building the 'removeItem' function, but even trying to fake the test out I couldn't get the particular assertion to pass. I believe there may be a typo., but for now, commenting it out might be an easy way to not frustrate a student for too long. Maybe I'm wrong though, I would love to be shown a solution (besides the 2-3 that I came up with) that passes the test.